errores

+++ admin/subirPersonal.php
@@ -9,12 +9,11 @@ if(isset($_POST) and $_SERVER["REQUEST_METHOD"]=="POST"){
 
     extract($_POST);
                 if($nombreprofesor!=""){
-                    
+                    echo "Entro al primer if";
                     //agregar archivos al sitioWeb
                     $targetfolder = "CV/";
                     $targetfoto = "foto/";
                     $targetfolder = $targetfolder . basename( $_FILES['archivo']['name']) ;
-                    $targetfoto = $targetfoto . basename( $_FILES['archivoFoto']['name']) ;
                    if(move_uploaded_file($_FILES['archivo']['tmp_name'], $targetfolder))
                     {
                         //echo "The file ". basename( $_FILES['archivo']['name']). " is uploaded";
@@ -22,23 +21,28 @@ if(isset($_POST) and $_SERVER["REQUEST_METHOD"]=="POST"){
                     else {
                         $response_array['status'] = 'errorArchivo';
                         echo json_encode($response_array);
-                        exit;
-                         
+                        exit;     
                     }
-                    if(move_uploaded_file($_FILES['archivoFoto']['tmp_name'], $targetfoto))
-                    {
-                        //echo "The file ". basename( $_FILES['archivo']['name']). " is uploaded";
+                    if(!file_exists($_FILES['archivoFoto']['tmp_name']) || !is_uploaded_file($_FILES['name']['tmp_name'])){
+                        $targetfoto = $targetfoto . basename( $_FILES['archivoFoto']['name']) ;
+                        if(move_uploaded_file($_FILES['archivoFoto']['tmp_name'], $targetfoto))
+                        {
+                            //echo "The file ". basename( $_FILES['archivo']['name']). " is uploaded";
+                        }
+                        else {
+                            $response_array['status'] = 'errorFoto';
+                            echo json_encode($response_array);
+                            exit;
+                        }
                     }
-                    else {
-                        $response_array['status'] = 'errorFoto';
-                        echo json_encode($response_array);
-                        exit;
+                    else{
+                        $targetfoto= "foto/default.jpg";
                     }
+                    
                     $archivo = $targetfolder;
                     $foto = $targetfoto;
                     $conexion = new DB();
                     $resultado = $conexion->insertarProfesor($nombreprofesor, $puesto, $carrera, $archivo, $foto);
-                    session_start();
                     //$_SESSION['result'] = 'guardado';
                     if($resultado>0){
                         $response_array['status'] = 'success';
@@ -54,6 +58,7 @@ if(isset($_POST) and $_SERVER["REQUEST_METHOD"]=="POST"){
 }
 else{
     
-    echo "se genero un error";
+    $response_array['status'] = 'errorConsulta';
+    echo json_encode($response_array);
 }
 ?>
